added sweet alert 2

+++ src/Pages/Myservices.jsx
@@ -2,6 +2,7 @@ import React, { use, useEffect, useState } from 'react';
 import { AuthContext } from '../Provider/AuthProvider';
 import { Link } from 'react-router';
 import axios from 'axios';
+import Swal from 'sweetalert2';
 
 const Myservices = () => {
     const { user } = use(AuthContext)
@@ -15,10 +16,31 @@ const Myservices = () => {
     }, [user?.email]);
 
     const handleDelete = (id) => {
-        axios.delete(`http://localhost:3000/delete/${id}`)
-            .then(res => console.log(res));
-        const filter = myServices.filter(service => service._id != id)
-        setMyServices(filter)
+       
+
+
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                axios.delete(`http://localhost:3000/delete/${id}`)
+                    .then(res => console.log(res));
+                const filter = myServices.filter(service => service._id != id)
+                setMyServices(filter)
+                Swal.fire({
+                    title: "Deleted!",
+                    text: "Your file has been deleted.",
+                    icon: "success"
+                });
+            }
+        });
+
     }
     return (
         <div>
